+++ app/controllers/handovers_controller.rb
@@ -1,11 +1,18 @@
 # frozen_string_literal: true
 
 class HandoversController < PrisonsApplicationController
+  include Sorting
+
   before_action :ensure_spo_user
 
   def index
     @pending_handover_count = @current_user.allocations.count(&:approaching_handover?)
-    @summary = SummaryService.new(:handovers, @prison, params['sort'])
-    @offenders = Kaminari.paginate_array(@summary.offenders.to_a).page(page)
+    offender_list = @prison.offenders.select(&:approaching_handover?)
+    allocations = @prison.allocations.where(nomis_offender_id: offender_list.map(&:offender_no))
+    offenders_with_allocs = offender_list.map do |o|
+      OffenderWithAllocationPresenter.new(o, allocations.detect { |a| a.nomis_offender_id == o.offender_no })
+    end
+    offenders = sort_collection offenders_with_allocs, default_sort: :last_name
+    @offenders = Kaminari.paginate_array(offenders).page(page)
   end
 end
