+++ app/views/debugging/prison_info.html.erb
@@ -29,7 +29,7 @@
 
   <div class="govuk-grid-column-one-quarter">
     <p class="govuk-data-value">
-      <%=  @summary.pending.count %>
+      <%= @summary.missing_info.count %>
     </p>
     <p class="govuk-data-label">
       With missing information
