+++ app/views/prisoners/allocated.html.erb
