+++ app/views/prisoners/missing_information.html.erb
@@ -1,0 +1,5 @@
+<% if @prison.womens? %>
+<%= render 'female_missing_information' %>
+<% else %>
+<%= render 'male_missing_information' %>
+<% end %>
