+++ app/views/prisoners/new_arrivals.html.erb
