+++ spec/spec_helper.rb
@@ -24,8 +24,8 @@ SimpleCov.start 'rails' do
   add_group "Services", "app/services"
 
   # Try to set this to current coverage levels so that it never goes down after a PR
-  # 13 lines uncovered at 99.63% coverage
-  minimum_coverage 99.63
+  # 11 lines uncovered at 99.67% coverage
+  minimum_coverage 99.67
   # sometimes coverage drops between branches - don't fail in these cases
   maximum_coverage_drop 0.5
